Begin CPUState refactor, fixed unit test failures
CPUControlSection is now run before it is unit tested.
CPUData section moved to its own file.
